+++ crates/module-system/hyperlane/tests/integration/with_agent/helpers/mod.rs
@@ -18,6 +18,7 @@ use crate::with_agent::helpers::evm::{
 use futures::future::join_all;
 use futures::{FutureExt, StreamExt};
 use sov_bank::Amount;
+use sov_hyperlane_integration::igp::IGPConfig;
 use sov_hyperlane_integration::EthAddress;
 use sov_modules_api::{CryptoSpec, HexHash, HexString, Spec};
 use sov_sequencer::preferred::PreferredSequencerConfig;
@@ -134,8 +135,11 @@ pub fn generate_setup() -> Setup {
     let sequencer = genesis_config.initial_sequencer.clone();
     let prover = genesis_config.initial_prover.clone();
 
+    let igp_config = IGPConfig {
+        admin: relayer.address(),
+    };
     let genesis_config =
-        GenesisConfig::from_minimal_config(genesis_config.into(), (), (), (), (), (), ());
+        GenesisConfig::from_minimal_config(genesis_config.into(), (), (), igp_config, (), (), ());
 
     Setup {
         sequencer,