+++ crates/module-system/module-implementations/extern/hyperlane-solana-register/tests/integration/setup.rs
@@ -2,6 +2,7 @@ use std::str::FromStr;
 use std::sync::Arc;
 
 use sov_bank::Amount;
+use sov_hyperlane_integration::igp::IGPConfig;
 use sov_hyperlane_integration::warp::{Admin, TokenKind};
 use sov_hyperlane_integration::{
     HyperlaneAddress, InterchainGasPaymaster, Ism, Mailbox as RawMailbox, MerkleTreeHook, Message,
@@ -77,12 +78,15 @@ pub fn setup() -> SetupParams {
         ism: Some(Ism::AlwaysTrust),
     };
 
+    let igp_config = IGPConfig {
+        admin: module_admin.address(),
+    };
     let genesis = GenesisConfig::from_minimal_config(
         genesis_config.clone().into(),
         (),
         (),
         (),
-        (),
+        igp_config,
         registration_conf,
     );